refactor: move config model to a directory
