+++ ui/projects/streampipes/platform-services/src/lib/model/config/email-config.model.ts
