+++ ui/projects/streampipes/platform-services/src/lib/model/config/general-config.model.ts
